toolkit/components/passwordmgr/test/mochitest/test_master_password.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for master password</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <script type="text/javascript" src="../../../prompts/test/prompt_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
Login Manager test: master password.

<p id="display"></p>

<div id="content" style="display: none">
<iframe id="iframe1"></iframe>
<iframe id="iframe2"></iframe>
</div>

<pre id="test">
<script class="testbody" type="text/javascript">
"use strict";

// Force parent to not look for tab-modal prompts, as they're not used for auth prompts.
isTabModal = false;

var exampleCom = "https://example.com/tests/toolkit/components/passwordmgr/test/mochitest/";
var exampleOrg = "https://example.org/tests/toolkit/components/passwordmgr/test/mochitest/";

gTestDependsOnDeprecatedLogin = true;
var chromeScript = runChecksAfterCommonInit();

runInParent(() => {
  const {Services} = ChromeUtils.import("resource://gre/modules/Services.jsm");

  var nsLoginInfo = new Components.Constructor("@mozilla.org/login-manager/loginInfo;1", Ci.nsILoginInfo);

  var login1 = new nsLoginInfo();
  var login2 = new nsLoginInfo();

  login1.init("https://example.com", "https://example.com", null,
              "user1", "pass1", "uname", "pword");
  login2.init("https://example.org", "https://example.org", null,
              "user2", "pass2", "uname", "pword");

  Services.logins.addLogin(login1);
  Services.logins.addLogin(login2);
});

var iframe1 = document.getElementById("iframe1");
var iframe2 = document.getElementById("iframe2");

add_task(async function test_1() {
  ok(isLoggedIn(), "should be initially logged in (no MP)");
  enableMasterPassword();
  ok(!isLoggedIn(), "should be logged out after setting MP");

  // Trigger a MP prompt via the API
  var state = {
    msg: "Please enter your master password.",
    title: "the title",
    textValue: "",
    passValue: "",
    iconClass: "authentication-icon question-icon",
    titleHidden: true,
    textHidden: true,
    passHidden: false,
    checkHidden: true,
    checkMsg: "",
    checked: false,
    focused: "passField",
    defButton: "button0",
  };
  var action = {
    buttonClick: "ok",
    passField: MASTER_PASSWORD,
  };
  var promptDone = handlePrompt(state, action);

  var logins = LoginManager.getAllLogins();

  await promptDone;
  is(logins.length, 3, "expected number of logins");

  ok(isLoggedIn(), "should be logged in after MP prompt");
  logoutMasterPassword();
  ok(!isLoggedIn(), "should be logged out");
});

add_task(async function test_2() {
  // Try again but click cancel.
  var state = {
    msg: "Please enter your master password.",
    title: "the title",
    textValue: "",
    passValue: "",
    iconClass: "authentication-icon question-icon",
    titleHidden: true,
    textHidden: true,
    passHidden: false,
    checkHidden: true,
    checkMsg: "",
    checked: false,
    focused: "passField",
    defButton: "button0",
  };
  var action = {
    buttonClick: "cancel",
  };
  var promptDone = handlePrompt(state, action);

  var failedAsExpected = false;
  var logins = null;
  try {
    logins = LoginManager.getAllLogins();
  } catch (e) {
    failedAsExpected = true;
  }
  await promptDone;
  ok(failedAsExpected, "getAllLogins should have thrown");
  is(logins, null, "shouldn't have gotten logins");
  ok(!isLoggedIn(), "should still be logged out");
});

add_task(async function test_3() {
  var state = {
    msg: "Please enter your master password.",
    title: "the title",
    textValue: "",
    passValue: "",
    iconClass: "authentication-icon question-icon",
    titleHidden: true,
    textHidden: true,
    passHidden: false,
    checkHidden: true,
    checkMsg: "",
    checked: false,
    focused: "passField",
    defButton: "button0",
  };
  var action = {
    buttonClick: "ok",
    passField: MASTER_PASSWORD,
  };
  var promptDone = handlePrompt(state, action);

  var fillPromise = promiseFormsProcessed();

  info("Load a single iframe to trigger a MP");
  iframe1.src = exampleCom + "subtst_master_pass.html";

  await promptDone;
  info("promptDone");
  await fillPromise;
  info("filled");

  // check contents of iframe1 fields
  var u = SpecialPowers.wrap(iframe1).contentDocument.getElementById("userfield");
  var p = SpecialPowers.wrap(iframe1).contentDocument.getElementById("passfield");
  is(u.value, "user1", "checking expected user to have been filled in");
  is(p.value, "pass1", "checking expected pass to have been filled in");
  info("clearing fields to not cause a submission when the next document is loaded");
  u.value = "";
  p.value = "";

  ok(isLoggedIn(), "should be logged in");
  logoutMasterPassword();
  ok(!isLoggedIn(), "should be logged out");
});

add_task(async function test_4() {
  var state = {
    msg: "Please enter your master password.",
    title: "the title",
    textValue: "",
    passValue: "",
    iconClass: "authentication-icon question-icon",
    titleHidden: true,
    textHidden: true,
    passHidden: false,
    checkHidden: true,
    checkMsg: "",
    checked: false,
    focused: "passField",
    defButton: "button0",
  };
  var action = {
    buttonClick: "none",
  };
  var promptDone = handlePrompt(state, action);

  // first part of loading 2 MP-triggering iframes
  iframe1.src = exampleOrg + "subtst_master_pass.html";
  // The MP prompt is open but don't take any action yet.
  await promptDone;

  // check contents of iframe1 fields
  var u = SpecialPowers.wrap(iframe1).contentDocument.getElementById("userfield");
  var p = SpecialPowers.wrap(iframe1).contentDocument.getElementById("passfield");
  is(u.value, "", "checking expected empty user");
  is(p.value, "", "checking expected empty pass");

  ok(!isLoggedIn(), "should be logged out");

  // XXX check that there's 1 MP window open

  // Load another iframe with a login form
  // This should detect that there's already a pending MP prompt, and not
  // put up a second one.
  var loadPromise = new Promise(resolve => {
    iframe2.addEventListener("load", function onload() {
      resolve();
    }, { once: true });
  });

  iframe2.src = exampleCom + "subtst_master_pass.html";
  await loadPromise;

  await new Promise(resolve => {
    // Testing a negative, wait a little to give the login manager a chance to
    // (incorrectly) fill in the form.  Note, we cannot use setTimeout()
    // here because the modal window suspends all window timers.  Instead we
    // must use a chrome script to use nsITimer directly.
    let chromeURL = SimpleTest.getTestFileURL("chrome_timeout.js");
    let script = SpecialPowers.loadChromeScript(chromeURL);
    script.addMessageListener("ready", _ => {
      script.sendAsyncMessage("setTimeout", { delay: 500 });
    });
    script.addMessageListener("timeout", resolve);
  });

  // iframe2 should load without having triggered a MP prompt (because one
  // is already waiting)

  // check contents of iframe2 fields
  u = SpecialPowers.wrap(iframe2).contentDocument.getElementById("userfield");
  p = SpecialPowers.wrap(iframe2).contentDocument.getElementById("passfield");
  is(u.value, "", "checking expected empty user");
  is(p.value, "", "checking expected empty pass");

  // XXX check that there's 1 MP window open
  ok(!isLoggedIn(), "should be logged out");

  // Ok, now enter the MP. The MP prompt is already up.

  var fillPromise = promiseFormsProcessed(2);

  // fill existing MP dialog with MP.
  action = {
    buttonClick: "ok",
    passField: MASTER_PASSWORD,
  };
  await handlePrompt(state, action);
  await fillPromise;

  // We shouldn't have to worry about iframe1's load event racing with
  // filling of iframe2's data. We notify observers synchronously, so
  // iframe2's observer will process iframe2 before iframe1 even finishes
  // processing the form.
  ok(isLoggedIn(), "should be logged in");

  // check contents of iframe1 fields
  u = SpecialPowers.wrap(iframe1).contentDocument.getElementById("userfield");
  p = SpecialPowers.wrap(iframe1).contentDocument.getElementById("passfield");
  is(u.value, "user2", "checking expected user to have been filled in");
  is(p.value, "pass2", "checking expected pass to have been filled in");

  info("clearing fields to not cause a submission when the next document is loaded");
  u.value = "";
  p.value = "";

  // check contents of iframe2 fields
  u = SpecialPowers.wrap(iframe2).contentDocument.getElementById("userfield");
  p = SpecialPowers.wrap(iframe2).contentDocument.getElementById("passfield");
  is(u.value, "user1", "checking expected user to have been filled in");
  is(p.value, "pass1", "checking expected pass to have been filled in");

  info("clearing fields to not cause a submission when the next document is loaded");
  u.value = "";
  p.value = "";
});

// XXX do a test5ABC with clicking cancel?

SimpleTest.registerCleanupFunction(function finishTest() {
  disableMasterPassword();
});

</script>
</pre>
</body>
</html>
